Set the request_uri to the request’s referrer.
This commit also organizes test/app/lifecycle.coffee a bit more with
`context`s and `describes`.

Before this change, the redirect_uri only matched on `redirect-to`-like
indicators in the request query string or payload.

This should *not* be a breaking change to any non-AJAX clients because
we return the `request.referral` _only if_ there are no other
`redirect-to`-like indicators.

However, to handle the case where we want auth with Gravity from an AJAX
request, the redirect_uri that we would want to expose to Gravity is the
referrer of the request. Example data flow:

1. GET www.artsy.net/auctions

(User opens modal and submits valid credentials, which leads to)

2. [client-side] POST /log_in

(Force’s server via artsy-passport returns a JSON payload which includes
api_sign_in_with_redirect URL, which leads to)

N.B: this is the stage where we the browser get’s a session cookie for
Force

3. [client-side] GET api.artsy.net/users/sign_in?redirectUri=www.artsy.net/auctions&trust_token=some-trust-token

(Gravity’s server handles the auth’ed get request and issues a redirect
to)

4. GET www.artsy.net/auctions

(We’re back were we started, but the browser has visited Gravity and
authenticated against it)

Jira: https://artsyproduct.atlassian.net/browse/AUCT-281